+++ plugin_ide.ui/src-lang/melnorme/lang/ide/ui/editor/actions/AbstractEditorHandler.java
@@ -13,12 +13,15 @@ package melnorme.lang.ide.ui.editor.actions;
 
 import static melnorme.utilbox.core.Assert.AssertNamespace.assertNotNull;
 import melnorme.lang.ide.ui.EditorSettings_Actual;
+import melnorme.lang.ide.ui.LangUIMessages;
+import melnorme.lang.ide.ui.actions.UIUserInteractionsHelper;
 import melnorme.lang.ide.ui.editor.AbstractLangEditor;
 
 import org.eclipse.core.commands.AbstractHandler;
 import org.eclipse.core.commands.ExecutionEvent;
 import org.eclipse.core.commands.ExecutionException;
 import org.eclipse.swt.widgets.Shell;
+import org.eclipse.ui.IEditorPart;
 import org.eclipse.ui.IWorkbenchPage;
 import org.eclipse.ui.editors.text.TextEditorActionContributor;
 import org.eclipse.ui.handlers.HandlerUtil;
@@ -37,8 +40,18 @@ public abstract class AbstractEditorHandler extends AbstractHandler {
 		this.page = assertNotNull(page);
 	}
 	
+	protected abstract String getOperationName();
+	
 	protected Shell getShell() {
 		return page.getWorkbenchWindow().getShell();
+	}
+	
+	protected ITextEditor getTextEditor() {
+		IEditorPart activeEditor = page.getActiveEditor();
+		if(activeEditor instanceof ITextEditor) {
+			return (ITextEditor) activeEditor;
+		}
+		return null;
 	}
 	
 	@Override
@@ -51,16 +64,29 @@ public abstract class AbstractEditorHandler extends AbstractHandler {
 		ITextEditor editor = (ITextEditor) HandlerUtil.getActiveEditorChecked(event);
 		
 		new LangEditorRunner(getShell(), editor) {
-			
 			@Override
 			protected void runWithLangEditor(AbstractLangEditor editor) {
 				doRunWithEditor(editor);
 			}
+			
+			@Override
+			protected void handleInternalError(String message) {
+				AbstractEditorHandler.this.handleInternalError(message);
+			};
 		}.run();
 		
 		return null;
 	}
 	
+	protected void handleInternalError(String message) {
+		handleError(LangUIMessages.InternalError + ": " + message);
+	}
+	
+	protected void handleError(String message) {
+		UIUserInteractionsHelper.openError(getShell(), getOperationName(), message);
+	}
+	
+	/** editor: not null */
 	protected abstract void doRunWithEditor(AbstractLangEditor editor);
 	
 }