+++ plugin_ide.ui/src/melnorme/lang/ide/ui/LangUIMessages_Actual.java
@@ -1,7 +1,7 @@
 package melnorme.lang.ide.ui;
 
 
-public abstract class LangUIMessages_Actual {
+public abstract class LangUIMessages_Actual implements LangUIMessages_Default {
 	
 	public static String InitializeAfterLoadJob_starter_job_name
 		= "Starting LANGUAGE UI Plugin initialization";